<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #41480

Problem Summary:

The [distsql concurrency of limit statement is set to 1](https://github.com/pingcap/tidb/blob/f9e1845849bf2b1f9f2e35cd0f5a5c8e397f2405/distsql/request_builder.go#L159-L161) before, but with partitioned table, it may affect performance, see the case in #41480.

```sql
MySQL [test]> CREATE TABLE t(id int PRIMARY KEY, val int)
    -> PARTITION BY RANGE (id)
    -> (PARTITION p1 VALUES LESS THAN (100),
    ->  PARTITION p2 VALUES LESS THAN (200),
    ->  PARTITION p3 VALUES LESS THAN (300),
    ->  PARTITION p4 VALUES LESS THAN (400),
    ->  PARTITION p5 VALUES LESS THAN (500),
    ->  PARTITION p6 VALUES LESS THAN (600),
    ->  PARTITION p7 VALUES LESS THAN (700),
    ->  PARTITION p8 VALUES LESS THAN (800),
    ->  PARTITION p9 VALUES LESS THAN (900),
    ->  PARTITION p10 VALUES LESS THAN (1000),
    ->  PARTITION p11 VALUES LESS THAN (1100),
    ->  PARTITION p12 VALUES LESS THAN (1200),
    ->  PARTITION p13 VALUES LESS THAN (1300),
    ->  PARTITION p14 VALUES LESS THAN (1400),
    ->  PARTITION p15 VALUES LESS THAN (1500));
Query OK, 0 rows affected (0.134 sec)

MySQL [test]> 
MySQL [test]> INSERT INTO t VALUES(50, 50), (150, 150), (250, 250);
Query OK, 3 rows affected (0.004 sec)
Records: 3  Duplicates: 0  Warnings: 0

MySQL [test]> ANALYZE TABLE t;
Query OK, 0 rows affected, 30 warnings (1.414 sec)

## Before this PR
MySQL [test]> EXPLAIN ANALYZE SELECT * FROM t ORDER BY id ASC LIMIT 1;
+----------------------------+---------+---------+-----------+---------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
| id                         | estRows | actRows | task      | access object | execution info                                                                                                                                                                                                                                   | operator info                | memory    | disk |
+----------------------------+---------+---------+-----------+---------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
| TopN_7                     | 1.00    | 1       | root      |               | time:10.6ms, loops:2                                                                                                                                                                                                                             | test.t.id, offset:0, count:1 | 280 Bytes | N/A  |
| └─TableReader_14           | 1.00    | 3       | root      | partition:all | time:10.6ms, loops:3, cop_task: {num: 18, max: 869.7µs, min: 423.9µs, avg: 536.6µs, p95: 869.7µs, max_proc_keys: 1, p95_proc_keys: 1, rpc_num: 18, rpc_time: 9.25ms, copr_cache_hit_ratio: 0.00, distsql_concurrency: 1}                         | data:Limit_13                | 319 Bytes | N/A  |
|   └─Limit_13               | 1.00    | 3       | cop[tikv] |               | tikv_task:{proc max:1ms, min:0s, avg: 55.6µs, p80:0s, p95:1ms, iters:18, tasks:18}, scan_detail: {total_process_keys: 3, total_process_keys_size: 113, total_keys: 21, get_snapshot_time: 495.3µs, rocksdb: {key_skipped_count: 3, block: {}}}   | offset:0, count:1            | N/A       | N/A  |
|     └─TableFullScan_12     | 3.00    | 3       | cop[tikv] | table:t       | tikv_task:{proc max:1ms, min:0s, avg: 55.6µs, p80:0s, p95:1ms, iters:18, tasks:18}                                                                                                                                                               | keep order:false             | N/A       | N/A  |
+----------------------------+---------+---------+-----------+---------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
4 rows in set (0.012 sec)

## After this PR
MySQL [test]> EXPLAIN ANALYZE SELECT * FROM t ORDER BY id ASC LIMIT 1;
+----------------------------+---------+---------+-----------+---------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
| id                         | estRows | actRows | task      | access object | execution info                                                                                                                                                                                                                            | operator info                | memory    | disk |
+----------------------------+---------+---------+-----------+---------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
| TopN_7                     | 1.00    | 1       | root      |               | time:1.47ms, loops:2                                                                                                                                                                                                                      | test.t.id, offset:0, count:1 | 280 Bytes | N/A  |
| └─TableReader_14           | 1.00    | 3       | root      | partition:all | time:1.43ms, loops:3, cop_task: {num: 15, max: 1.11ms, min: 891.7µs, avg: 982.6µs, p95: 1.11ms, max_proc_keys: 1, p95_proc_keys: 1, rpc_num: 15, rpc_time: 13.9ms, copr_cache_hit_ratio: 0.00, distsql_concurrency: 15}                   | data:Limit_13                | 2.92 KB   | N/A  |
|   └─Limit_13               | 1.00    | 3       | cop[tikv] |               | tikv_task:{proc max:0s, min:0s, avg: 0s, p80:0s, p95:0s, iters:15, tasks:15}, scan_detail: {total_process_keys: 3, total_process_keys_size: 113, total_keys: 18, get_snapshot_time: 654.6µs, rocksdb: {key_skipped_count: 3, block: {}}}  | offset:0, count:1            | N/A       | N/A  |
|     └─TableFullScan_12     | 3.00    | 3       | cop[tikv] | table:t       | tikv_task:{proc max:0s, min:0s, avg: 0s, p80:0s, p95:0s, iters:15, tasks:15}                                                                                                                                                              | keep order:true              | N/A       | N/A  |
+----------------------------+---------+---------+-----------+---------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------+-----------+------+
4 rows in set (0.003 sec)
```

### What is changed and how it works?

Set the concurrency to the number of partitions, with a upper limit of distsql concurrency(default to 15).

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
store/copr: set concurrency to partition number for limit statement to reduce latency.
```